Moved pseudo-instructions into its own function. Added an error to each pseudo-instruction for if there are not the right number of operands.
