+++ src/tests/parser.rs
@@ -1,3 +1,4 @@
 pub mod assembling;
 pub mod parser_assembler_main;
 pub mod parsing;
+pub mod pseudo_instruction_parsing;
